agno-agent/pyproject.toml
[project]
name = "agno-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.6",
    "instructor>=1.10.0",
    "mcp>=1.12.3",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
]
